+++ core/src/main/java/org/testcontainers/containers/ComposeServiceWaitStrategyTarget.java
@@ -5,7 +5,6 @@ import com.github.dockerjava.api.model.Container;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NonNull;
-import org.slf4j.Logger;
 import org.testcontainers.DockerClientFactory;
 import org.testcontainers.containers.wait.strategy.WaitStrategyTarget;
 
@@ -23,23 +22,24 @@ class ComposeServiceWaitStrategyTarget implements WaitStrategyTarget {
 
     private final Container container;
     private final GenericContainer proxyContainer;
-    @Getter
-    private final Logger logger;
     @NonNull
     private Map<Integer, Integer> mappedPorts;
-    @Getter
-    private List<Integer> exposedPorts = new ArrayList<>();
     @Getter(lazy=true)
     private final InspectContainerResponse containerInfo = DockerClientFactory.instance().client().inspectContainerCmd(getContainerId()).exec();
 
     ComposeServiceWaitStrategyTarget(Container container, GenericContainer proxyContainer,
-                                     Logger logger, @NonNull Map<Integer, Integer> mappedPorts) {
+                                     @NonNull Map<Integer, Integer> mappedPorts) {
         this.container = container;
-
         this.proxyContainer = proxyContainer;
-        this.logger = logger;
         this.mappedPorts = new HashMap<>(mappedPorts);
-        this.exposedPorts.addAll(this.mappedPorts.keySet());
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<Integer> getExposedPorts() {
+        return new ArrayList<>(this.mappedPorts.keySet());
     }
 
     /**