+++ core/src/main/java/org/testcontainers/containers/wait/strategy/HostPortWaitStrategy.java
@@ -25,7 +25,7 @@ public class HostPortWaitStrategy extends AbstractWaitStrategy {
     protected void waitUntilReady() {
         final Set<Integer> externalLivenessCheckPorts = getLivenessCheckPorts();
         if (externalLivenessCheckPorts.isEmpty()) {
-            log.debug("Liveness check ports of {} is empty. Not waiting.", waitStrategyTarget.getContainerName());
+            log.debug("Liveness check ports of {} is empty. Not waiting.", waitStrategyTarget.getContainerInfo().getName());
             return;
         }