+++ core/src/test/java/org/testcontainers/junit/DockerComposeWaitStrategyTest.java
@@ -2,7 +2,6 @@ package org.testcontainers.junit;
 
 import org.junit.Test;
 import org.junit.runner.Description;
-import org.rnorth.ducttape.TimeoutException;
 import org.rnorth.visibleassertions.VisibleAssertions;
 import org.testcontainers.containers.DockerComposeContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
@@ -22,7 +21,7 @@ public class DockerComposeWaitStrategyTest {
         try {
             environment.starting(Description.createTestDescription(Object.class, "name"));
             VisibleAssertions.pass("Docker compose should start after waiting for listening port");
-        } catch (TimeoutException e) {
+        } catch (RuntimeException e) {
             VisibleAssertions.fail("Docker compose should start after waiting for listening port with failed with: " + e);
         }
     }
@@ -37,7 +36,7 @@ public class DockerComposeWaitStrategyTest {
         try {
             environment.starting(Description.createTestDescription(Object.class, "name"));
             VisibleAssertions.pass("Docker compose should start after waiting for listening port");
-        } catch (TimeoutException e) {
+        } catch (RuntimeException e) {
             VisibleAssertions.fail("Docker compose should start after waiting for listening port with failed with: " + e);
         }
     }
@@ -45,23 +44,21 @@ public class DockerComposeWaitStrategyTest {
     @Test
     public void testWaitingFails() {
         final DockerComposeContainer environment = new DockerComposeContainer(new File("src/test/resources/compose-test.yml"))
-            .withStartupTimeout(Duration.ofSeconds(10))
-            .withExposedService("redis_1", REDIS_PORT, Wait.forHttp("/test"));
+            .withExposedService("redis_1", REDIS_PORT, Wait.forHttp("/test").withStartupTimeout(Duration.ofSeconds(10)));
         VisibleAssertions.assertThrows("waiting on an invalid http path times out",
-            TimeoutException.class,
+            RuntimeException.class,
             () -> environment.starting(Description.createTestDescription(Object.class, "name")));
     }
 
     @Test
     public void testWaitOnOneOfMultipleStrategiesFailing() {
         final DockerComposeContainer environment = new DockerComposeContainer(new File("src/test/resources/compose-test.yml"))
-            .withStartupTimeout(Duration.ofSeconds(10))
-            .withExposedService("redis_1", REDIS_PORT, Wait.forListeningPort())
-            .withExposedService("db_1", 3306, Wait.forLogMessage(".*test test test.*\\s", 1))
+            .withExposedService("redis_1", REDIS_PORT, Wait.forListeningPort().withStartupTimeout(Duration.ofSeconds(10)))
+            .waitingFor("db_1", Wait.forLogMessage(".*test test test.*\\s", 1).withStartupTimeout(Duration.ofSeconds(10)))
             .withTailChildContainers(true);
 
         VisibleAssertions.assertThrows("waiting on one failing strategy to time out",
-            TimeoutException.class,
+            RuntimeException.class,
             () -> environment.starting(Description.createTestDescription(Object.class, "name")));
     }